+++ tests/testthat/test-portion.R
@@ -100,17 +100,17 @@ test_that("portions works", {
 
   # portion_df
   expect_equal(
-    septic_patients %>% select(AMX) %>% portion_df() %>% pull(Value),
+    septic_patients %>% select(AMX) %>% portion_df() %>% pull(value),
     c(septic_patients$AMX %>% portion_SI(),
       septic_patients$AMX %>% portion_R())
   )
   expect_equal(
-    septic_patients %>% select(AMX) %>% portion_df(combine_IR = TRUE) %>% pull(Value),
+    septic_patients %>% select(AMX) %>% portion_df(combine_IR = TRUE) %>% pull(value),
     c(septic_patients$AMX %>% portion_S(),
       septic_patients$AMX %>% portion_IR())
   )
   expect_equal(
-    septic_patients %>% select(AMX) %>% portion_df(combine_SI = FALSE) %>% pull(Value),
+    septic_patients %>% select(AMX) %>% portion_df(combine_SI = FALSE) %>% pull(value),
     c(septic_patients$AMX %>% portion_S(),
       septic_patients$AMX %>% portion_I(),
       septic_patients$AMX %>% portion_R())